pkg/core/response.go
package core

import (
	"encoding/json"
	"net/http"
)

type Response struct {
	RawResponse 	*http.Response
	body       		[]byte
}

func (r *Response) UnmarshalJson(target interface{}) error {
	if len(r.body) > 0{
		return json.Unmarshal(r.body, target)
	}

	return nil
}

func (r *Response) Body() []byte {
	if r.RawResponse == nil {
		return []byte{}
	}
	return r.body
}

func (r *Response) Status() string {
	if r.RawResponse == nil {
		return ""
	}
	return r.RawResponse.Status
}

func (r *Response) StatusCode() int {
	if r.RawResponse == nil {
		return 0
	}
	return r.RawResponse.StatusCode
}

// IsSuccess method returns true if HTTP status `code >= 200 and <= 299` otherwise false.
func (r *Response) IsSuccess() bool {
	return r.StatusCode() > 199 && r.StatusCode() < 300
}

// IsError method returns true if HTTP status `code >= 400` otherwise false.
func (r *Response) IsError() bool {
	return r.StatusCode() > 399
}
